add(upseh2):start work with auth

+++ upseh2/src/main/java/com/example/upseh2/dtos/BalanceDTO.java
@@ -2,8 +2,11 @@ package com.example.upseh2.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
+
 @Data
 public class BalanceDTO {
     private long id;
-    private long balance;
+    private BigDecimal balance;
 }
